Feat #73 (#78)
* feat :: 어드민 통계 페이지 api 추가

* feat :: 신입생 명단 페이지 api 추가

* refactor :: excel 길이 설정 버그 수정

* refactor :: Query문 수정

* refactor :: 클래스 이름 변경

* feat :: 신입생 명단 테이블 추가

* feat :: 어드민 2차 전형 페이지 api추가 및 엑셀추가

* feat :: 어드민 1차 성적 페이지 api 추가

* Delete ida-http-application/src/main/resources/application.yml

* refactor :: 코드리뷰 반영

+++ ida-infrastructure/src/main/kotlin/com/dgswiphak/ida/persistence/secondary/entity/value/InterviewVO.kt
@@ -5,8 +5,8 @@ import jakarta.persistence.Embeddable
 @Embeddable
 class InterviewVO(
 
-    val studyCompetence: Int? = 0,
+    val studyCompetence: Double? = 0.0,
 
-    val computingCompetence: Int? = 0
+    val computingCompetence: Double? = 0.0
 
 )
